add aria label, color checkboxes and inactive category.
