+++ src/App.css
@@ -11,3 +11,66 @@ button,
 input {
   margin: 1em auto 0;
 }
+
+label.for-checkbox {
+  display: block;
+  cursor: pointer;
+}
+
+input[type='checkbox'] {
+  position: relative;
+  cursor: pointer;
+  width: 16px;
+  height: 16px;
+  margin: 0;
+  margin-right: 10px;
+}
+
+input[type='checkbox']::before {
+  content: '';
+  display: block;
+  position: absolute;
+  width: 16px;
+  height: 16px;
+  top: 0;
+  left: 0;
+  border: 1px solid #555555;
+  border-radius: 3px;
+  background-color: white;
+}
+
+input[type='checkbox']:checked:after {
+  content: '';
+  display: block;
+  width: 5px;
+  height: 10px;
+  border: solid black;
+  border-width: 0 2px 2px 0;
+  -webkit-transform: rotate(45deg);
+  -ms-transform: rotate(45deg);
+  transform: rotate(45deg);
+  position: absolute;
+  top: 2px;
+  left: 6px;
+}
+
+input[type='checkbox'].category-soon::before {
+  background-color: #ff968a;
+}
+
+input[type='checkbox'].category-kind-of-soon::before {
+  background-color: #ffffbf;
+}
+
+input[type='checkbox'].category-not-soon::before {
+  background-color: #cce2cb;
+}
+
+input[type='checkbox'].category-inactive::before,
+input[type='checkbox'].category-inactive::after {
+  background-color: #f2f2f2;
+}
+input[type='checkbox']:disabled::before,
+input[type='checkbox']:disabled::after {
+  background-color: #b6b6b6;
+}
